refactor!: make split for internal agent and server
